backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
authlib
httpx
itsdangerous
python-jose[cryptography]
pyjwt
cryptography
aiofiles
